# Optimize encode and decode with TypedArrays