--- a/src/sourcemap-codec.ts
+++ b/src/sourcemap-codec.ts
@@ -5,12 +5,30 @@ export type SourceMapSegment =
 export type SourceMapLine = SourceMapSegment[];
 export type SourceMapMappings = SourceMapLine[];
 
-const charToInteger: { [charCode: number]: number } = {};
 const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';
+const intToChar = new Uint8Array(65); // 65 possible chars.
+const charToInteger = new Uint8Array(123); // z is 122 in ASCII
 
+// Fill the buffer with 255, so we can detect invalid chars.
+for (let i = 0; i < 128; i++) charToInteger[i] = 255;
+// And fill with valid values.
 for (let i = 0; i < chars.length; i++) {
-	charToInteger[chars.charCodeAt(i)] = i;
+	const c = chars.charCodeAt(i);
+	charToInteger[c] = i;
+	intToChar[i] = c;
 }
+
+// Provide a fallback for older environments.
+const td = typeof TextDecoder !== 'undefined' ? new TextDecoder('ascii') : {
+	decode(buf: Uint8Array) {
+		let out = '';
+		for (let i = 0; i < buf.length; i++) {
+			out += String.fromCharCode(buf[i])
+		}
+		return out;
+	}
+};
+
 
 export function decode(mappings: string): SourceMapMappings {
 	const decoded: SourceMapMappings = [];
@@ -40,7 +58,7 @@ export function decode(mappings: string): SourceMapMappings {
 
 		} else {
 			let integer = charToInteger[c];
-			if (integer === undefined) {
+			if (integer === 255) {
 				throw new Error('Invalid character (' + String.fromCharCode(c) + ')');
 			}
 
@@ -90,26 +108,36 @@ export function encode(decoded: SourceMapMappings): string {
 	let sourceCodeLine = 0;   // third field
 	let sourceCodeColumn = 0; // fourth field
 	let nameIndex = 0;        // fifth field
-	let mappings = '';
+
+	let buf = new Uint8Array(1000);
+	let pos = 0;
 
 	for (let i = 0; i < decoded.length; i++) {
 		const line = decoded[i];
-		if (i > 0) mappings += ';';
+		if (i > 0) {
+			buf = reserve(buf, pos, 1);
+			buf[pos++] = 59; // ";"
+		}
 		if (line.length === 0) continue;
 
 		let generatedCodeColumn = 0; // first field
 
 		const lineMappings: string[] = [];
 
-		for (const segment of line) {
-			let segmentMappings = encodeInteger(segment[0] - generatedCodeColumn);
+		for (let j = 0; j < line.length; j++) {
+			const segment = line[j];
+			// We can push up to 5 ints, each int can take at most 7 chars, and we
+			// may push a comma.
+			buf = reserve(buf, pos, 36);
+			if (j > 0) buf[pos++] = 44; // ","
+
+			pos = encodeInteger(buf, pos, segment[0] - generatedCodeColumn);
 			generatedCodeColumn = segment[0];
 
 			if (segment.length > 1) {
-				segmentMappings +=
-					encodeInteger(segment[1] - sourceFileIndex) +
-					encodeInteger(segment[2] - sourceCodeLine) +
-					encodeInteger(segment[3] - sourceCodeColumn);
+				pos = encodeInteger(buf, pos, segment[1] - sourceFileIndex);
+				pos = encodeInteger(buf, pos, segment[2] - sourceCodeLine);
+				pos = encodeInteger(buf, pos, segment[3] - sourceCodeColumn);
 
 				sourceFileIndex = segment[1];
 				sourceCodeLine = segment[2];
@@ -117,30 +145,32 @@ export function encode(decoded: SourceMapMappings): string {
 			}
 
 			if (segment.length === 5) {
-				segmentMappings += encodeInteger(segment[4] - nameIndex);
+				pos = encodeInteger(buf, pos, segment[4] - nameIndex);
 				nameIndex = segment[4];
 			}
-
-			lineMappings.push(segmentMappings);
 		}
-
-		mappings += lineMappings.join(',');
 	}
 
-	return mappings;
+	return td.decode(buf.subarray(0, pos));
 }
 
-function encodeInteger(num: number): string {
-	var result = '';
+function reserve(buf: Uint8Array, pos: number, count: number): Uint8Array {
+	if (buf.length <= pos + count) {
+		const swap = new Uint8Array(buf.length * 2);
+		swap.set(buf);
+		buf = swap;
+	}
+	return buf;
+}
+
+function encodeInteger(buf: Uint8Array, pos: number, num: number): number {
 	num = num < 0 ? (-num << 1) | 1 : num << 1;
 	do {
-		var clamped = num & 31;
+		let clamped = num & 31;
 		num >>>= 5;
-		if (num > 0) {
-			clamped |= 32;
-		}
-		result += chars[clamped];
+		if (num > 0) clamped |= 32;
+		buf[pos++] = intToChar[clamped];
 	} while (num > 0);
 
-	return result;
+	return pos;
 }