--- a/tsconfig.json
+++ b/tsconfig.json
@@ -5,7 +5,7 @@
 		"noImplicitThis": true,
 		"noEmitOnError": true,
 		"target": "es5",
-		"lib": ["es5", "es6"],
+		"lib": ["es5", "es6", "dom"],
 		"declaration": true,
 		"outDir": "dist/types"
 	},