@jridgewell 2022-01-31
Friendly ping @Rich-Harris, is there anything you'd like me to do with this PR?